- add translations to .gitignore - add translation decorators for opsamef -> keras, opsamef -> canopy - refactor canopy translation helper methods - remove unused helper methods in opsamef -> canopy translator
